Leave some tests on tap
